fix warnings when cross-compiling to win32

+++ src/c_lib/sound/wav.cpp
@@ -37,10 +37,10 @@ void print_wav_data(WavData* data)
         "byte rate: %d\n"
         "bits per sample: %hd\n"
         "duration (seconds): %0.2f\n"
-        "size (bytes):  %llu\n"
+        "size (bytes):  %lu\n"
         "in use: %d\n",
         data->format, data->channels, data->sample_rate, data->byte_rate,
-        data->bits_per_sample, data->duration, (long long unsigned)data->size,
+        data->bits_per_sample, data->duration, (long unsigned)data->size,
         data->in_use);
 }
 
@@ -254,7 +254,7 @@ bool read_wav_data(FILE* f, WavData* data, unsigned char** buffer)
     read = fread(*buffer, sizeof(char), data->size, f);
     if (read != data->size)
     {
-        printf("Failed to read PCM data of size %llu\n", (long long unsigned)data->size);
+        printf("Failed to read PCM data of size %lu\n", (long unsigned)data->size);
         return false;
     }
