+++ src/net_lib/host.cpp
@@ -495,11 +495,11 @@ static void client_connect(ENetEvent* event)
         uint8_t address[4];
         address_from_uint32(event->peer->address.host, address);
 
-        printf("Client %d connected from %d.%d.%d.%d:%d. %llu clients connected\n",
+        printf("Client %d connected from %d.%d.%d.%d:%d. %lu clients connected\n",
                npm->client_id,
                address[0], address[1], address[2], address[3],
                event->peer->address.port,
-               (long long unsigned)NetServer::number_of_clients);
+               (long unsigned)NetServer::number_of_clients);
 
         Session* session = begin_session(event->peer->address.host, npm->client_id);
         users->assign_session_to_user(session);
@@ -526,7 +526,7 @@ static void client_disconnect(ENetPeer* peer, enet_uint32 data)
     NetPeer* nc = (NetPeer*)peer->data;
 
     ClientID client_id = NULL_CLIENT;
-    printf("Number of clients: %llu\n", (long long unsigned)NetServer::number_of_clients);
+    printf("Number of clients: %lu\n", (long unsigned)NetServer::number_of_clients);
     if (nc != NULL && nc->client_id >= 0 && nc->client_id < HARD_MAX_CONNECTIONS)
     {
         GS_ASSERT(NetServer::number_of_clients > 0);
@@ -581,7 +581,7 @@ static void client_disconnect(ENetPeer* peer, enet_uint32 data)
         else
         {
             GS_ASSERT(false);
-            printf("Client %d disconnected with unknown code %llu\n", client_id, (long long unsigned)data);
+            printf("Client %d disconnected with unknown code %lu\n", client_id, (long unsigned)data);
         }
     }
